backend/requirements.txt
Django
djangorestframework

markdown
django-filter

python-decouple
django-cors-headers
drf_spectacular
psycopg2-binary

Pillow

pytest
pytest-django
djangorestframework-simplejwt